test_network_nfs_io: Mounts will be readonly
We have decided to mount the NFS shares in the ptx-flavor readonly.
So this change inverts the logic and makes sure, that we can not write
onto the mounts.

And while we are on it: Do not assert during the iteration. This way we
can check all mounts and not only all until the first failure.