test_linux_hardware: Test for SPI devices instead of the subsystem
The previous test very bluntly only tested if the SPI-subsystem in the
kernel has been loaded at all. But this does not tell us anything about
the actual devices on the bus we want to use.

The new tests instead try to use the SPI devices the way userspace
would.

* The Ethernet switch is managed by DSA.
  By successfully reading the statistics we can assume that the SPI
  communication to the Ethernet Switch works.
* The ADC is managed by IIO.
  By successfully reading a non-zero ADC reading we can assume that
  we are reading actual values from the ADC.
* The LCD is managed by drm.
  We only check if the correct driver has been probed.