test/tad: Add test_tacd_dut_power_off_floating
This test makes sure the third state (On, Off, OffFloating) of the TACs
power switch works electrically.

This is especially important since this state is not exposed on the web
interface (or rather only in the API).

With this test we make sure that the load is actually engaged, when the
power switch is "Off" - and that we can also switch it to "OffFloating".